Move the definition of a date fieldset into its own component
